requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
pydantic==1.10.6
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
